common-package/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "common-lib"
version = "1.0.0"
description = "common package"
readme = "README.rst"
requires-python = ">=3.5"
license = { text = "MIT" }

authors = [
    { name = "N A Sifat", email = "mdnuraminsifat380@gmail.com" }
]

dependencies = [
    "Django==5.0.8",
    "django-cors-headers==4.4.0",
    "djangorestframework==3.15.2",
    "drf-yasg==1.21.7",
    "django-debug-toolbar==4.4.6",
    "django-environ==0.11.2",
    "django-extensions==3.2.3",
    "django-storages==1.14.4",
    "django-filter==24.2",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "psycopg2-binary==2.9.9",
    "aenum==3.1.15",
    "redis==4.0.2"
]

[project.optional-dependencies]
dev = [
    "setuptools>=64",
]
